#956: extract types out of file
