+++ src/blocks/combinators.ts
@@ -28,24 +28,22 @@ import {
   isReader,
   isRendererBlock,
   Logger,
-  UserOptions,
+  OutputKey,
   ReaderRoot,
   RenderedArgs,
   SanitizedServiceConfiguration,
   Schema,
   ServiceDependency,
-  TemplateEngine,
-  RegistryId,
-  OutputKey,
+  UserOptions,
 } from "@/core";
 import { validateInput, validateOutput } from "@/validators/generic";
 import {
   castArray,
+  fromPairs,
+  isEmpty,
   isPlainObject,
   mapValues,
   pickBy,
-  isEmpty,
-  fromPairs,
 } from "lodash";
 import { BusinessError, ContextError } from "@/errors";
 import {
@@ -65,54 +63,7 @@ import {
   PipelineConfigurationError,
 } from "@/blocks/errors";
 import { engineRenderer } from "@/utils/renderers";
-
-export type ReaderConfig =
-  | RegistryId
-  // eslint-disable-next-line @typescript-eslint/consistent-indexed-object-style -- Record<> doesn't allow labelled keys
-  | { [key: string]: ReaderConfig }
-  | ReaderConfig[];
-
-export interface BlockConfig {
-  id: RegistryId;
-
-  /**
-   * (Optional) human-readable label for the step. Shown in the progress indicator
-   */
-  label?: string;
-
-  /**
-   * (Optional) indicate the step is being run in the interface
-   */
-  notifyProgress?: boolean;
-
-  onError?: {
-    alert?: boolean;
-  };
-
-  window?: "self" | "opener" | "target" | "broadcast";
-
-  outputKey?: string;
-
-  /**
-   * (Optional) condition expression written in templateEngine for deciding if the step should be run. If not
-   * provided, the step is run unconditionally.
-   */
-  if?: string | boolean | number;
-
-  /**
-   * (Optional) root selector for reader
-   */
-  root?: string;
-
-  /**
-   * (Optional) template language to use for rendering the if and config properties. Default is mustache
-   */
-  templateEngine?: TemplateEngine;
-
-  config: Record<string, unknown>;
-}
-
-export type BlockPipeline = BlockConfig[];
+import { BlockConfig, BlockPipeline, ReaderConfig } from "./types";
 
 /** Return block definitions for all blocks referenced in a pipeline */
 export async function blockList(
