+++ src/blocks/types.ts
@@ -15,7 +15,57 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import { RegistryId, TemplateEngine } from "@/core";
+
 export interface Availability {
   matchPatterns?: string | string[];
   selectors?: string | string[];
 }
+
+export type ReaderConfig =
+  | RegistryId
+  // eslint-disable-next-line @typescript-eslint/consistent-indexed-object-style -- Record<> doesn't allow labelled keys
+  | { [key: string]: ReaderConfig }
+  | ReaderConfig[];
+
+export interface BlockConfig {
+  id: RegistryId;
+
+  /**
+   * (Optional) human-readable label for the step. Shown in the progress indicator
+   */
+  label?: string;
+
+  /**
+   * (Optional) indicate the step is being run in the interface
+   */
+  notifyProgress?: boolean;
+
+  onError?: {
+    alert?: boolean;
+  };
+
+  window?: "self" | "opener" | "target" | "broadcast";
+
+  outputKey?: string;
+
+  /**
+   * (Optional) condition expression written in templateEngine for deciding if the step should be run. If not
+   * provided, the step is run unconditionally.
+   */
+  if?: string | boolean | number;
+
+  /**
+   * (Optional) root selector for reader
+   */
+  root?: string;
+
+  /**
+   * (Optional) template language to use for rendering the if and config properties. Default is mustache
+   */
+  templateEngine?: TemplateEngine;
+
+  config: Record<string, unknown>;
+}
+
+export type BlockPipeline = BlockConfig[];
